Trim mdlFile 00342990 return width

+++ src/Battle/btlVoice.c
@@ -9805,8 +9805,8 @@ void func_002ee640(u64 *param_1)
   int iVar1;
   u8 *puVar2;
   int iVar3;
-  u64 uVar4;
-  u64 uVar5;
+  u32 uVar4;
+  u32 uVar5;
   int iVar6;
   
   *(u32 *)(DAT_007ce3ec + 0xc) = *(u32 *)(DAT_007ce3ec + 0xc) & 0xfdffffff;
